+++ src/silverlib/game/SceneImg.java
@@ -1,5 +1,6 @@
 package silverlib.game;
 
+import silverlib.geo.Point;
 import silverlib.geo.Shape;
 import silverlib.geo.Fillable;
 import silverlib.img.Img;
@@ -9,6 +10,9 @@ import java.awt.Color;
 public class SceneImg extends Img {
   public SceneImg(int width, int height) {
     super(width, height, Color.WHITE);
+  }
+  public SceneImg(int width, int height, Color bg) {
+    super(width, height, bg);
   }
 
   public void add(Shape obj) {
@@ -21,5 +25,17 @@ public class SceneImg extends Img {
 
   public void placeSprite(GameSprite sprite, int x, int y) {
     sprite.drawOn(this, x, y);
+  }
+
+  public void overlayImg(Img img) {
+    Point center = new Point(this.width()/2, this.height()/2);
+
+    Point no = new Point(center.x()-(img.width()/2), center.y()-(img.height()/2));
+
+    for (Point p : Point.iterPoints(0,0, img.width(), img.height())) {
+      Point np = Point.adjustForOrigin(p, no);
+
+      this.set(np.x(), np.y(), img.get(p.x(), p.y()));
+    }
   }
 }